Move os to public/node
