+++ packages/cli-kit/src/index.ts
@@ -7,7 +7,6 @@ export * as file from './file.js'
 export * as git from './git.js'
 export * as http from './http.js'
 export * as npm from './npm.js'
-export * as os from './os.js'
 export * as output from './output.js'
 export * as path from './path.js'
 export * as plugins from './plugins.js'
