+++ packages/cli-kit/src/private/node/analytics.ts
@@ -3,7 +3,7 @@ import BaseCommand from '../../public/node/base-command.js'
 import {CommandContent} from '../../public/node/hooks/prerun.js'
 import * as metadata from '../../metadata.js'
 import {macAddress} from '../../environment/local.js'
-import {platformAndArch} from '../../os.js'
+import {platformAndArch} from '../../public/node/os.js'
 import {hashString} from '../../string.js'
 import * as environment from '../../environment.js'
 import {Interfaces} from '@oclif/core'
