+++ packages/cli-kit/src/public/node/analytics.test.ts
@@ -4,7 +4,7 @@ import {startAnalytics} from '../../private/node/analytics.js'
 import {hashString} from '../../string.js'
 import * as environment from '../../environment.js'
 import {join as joinPath, dirname} from '../../path.js'
-import * as os from '../../os.js'
+import * as os from './os.js'
 import {mockAndCaptureOutput} from '../../testing/output.js'
 import {getAppInfo} from '../../store.js'
 import constants from '../../constants.js'
