+++ packages/cli-kit/src/session/store.ts
@@ -1,6 +1,6 @@
 import {SessionSchema} from './schema.js'
 import constants from '../constants.js'
-import {platformAndArch} from '../os.js'
+import {platformAndArch} from '../public/node/os.js'
 import {store as secureStore, fetch as secureFetch, remove as secureRemove} from '../secure-store.js'
 import {content, debug} from '../output.js'
 import {getSession, removeSession, setSession, clearAllAppInfo} from '../store.js'
